tests/wpt/mozilla/tests/mozilla/iframe_contentDocument.html

<html>
<head>
<meta charset="utf8" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<title>Iframe contentDocument test.</title>
</head>
<body>
<iframe src="resources/iframe_contentDocument_inner.html" id="iframe"></iframe>
<script>
async_test(function() {
  var iframe = document.getElementById('iframe');
  iframe.onload = this.step_func_done(function() {
    assert_equals(iframe.contentDocument.getElementById('test').textContent, 'value');
    this.done();
  })
});
</script>
</body>
</html>
